pyproject.toml
[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"

[project]
name = "git-ssh-wrapper"
description = ""
requires-python = ">=3.7"
authors = [
    {name = "Scott Chase Waggener", email = "tidalpaladin@protonmail.com"},
]
dependencies = [
    "pyyaml",
]
readme = "README.md"
license = {text = "Apache"}
dynamic = ["version"]

[project.optional-dependencies]

[tool.autoflake]
remove-all-unused-imports = true
remove-unused-variables = true

[tool.autopep8]
max_line_length = 120
ignore = "E501,W6,E203"
in-place = true
recursive = true
aggressive = 3

[tool.black]
line-length = 120
target-version = ['py37']

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 119
lines_after_imports = 2
multi_line_output = 3
use_parentheses = false

[tool.pyright]

[tool.pdm]
version = { source = "scm" }

[tool.pdm.dev-dependencies]
test = [
  "pytest",
  "pytest-mock",
  "pytest-cov",
  "coverage",
]
quality = [
  "autoflake",
  "autopep8",
  "black",
  "flake8",
  "isort",
]

[tool.pytest.ini_options.markers]
ci_skip = "skip tests in CI environment"

# Sample script config
[project.scripts]
wrapped-ssh = "git_ssh_wrapper.__main__:entrypoint"
